django_project/store_cms/requirements/local.txt
# Local development dependencies go here
-r base.txt

django-extensions==1.7.9
Werkzeug==0.12.2
django-debug-toolbar==1.8
ipdb==0.10.3
